+++ dangerzone/isolation_provider/container.py
@@ -262,8 +262,8 @@ class Container(IsolationProvider):
         # Convert document to pixels
         command = [
             "/usr/bin/python3",
-            "/usr/local/bin/dangerzone.py",
-            "document-to-pixels",
+            "-m",
+            "dangerzone.conversion.doc_to_pixels",
         ]
         extra_args = [
             "-v",
@@ -282,8 +282,8 @@ class Container(IsolationProvider):
             # Convert pixels to safe PDF
             command = [
                 "/usr/bin/python3",
-                "/usr/local/bin/dangerzone.py",
-                "pixels-to-pdf",
+                "-m",
+                "dangerzone.conversion.pixels_to_pdf",
             ]
             extra_args = [
                 "-v",
