+++ install/linux/build-image.sh
@@ -5,7 +5,7 @@ set -e
 TAG=dangerzone.rocks/dangerzone:latest
 
 echo "Building container image"
-podman build container --tag $TAG
+podman build dangerzone/ -f Dockerfile --tag $TAG
 
 echo "Saving and compressing container image"
 podman save $TAG | gzip > share/container.tar.gz
